Allow default values to be an empty string.
Previously, because of the way reflect.Tag.GetTag works,
a default value of an empty string would be interpreted
as a default value not being provided at all. As a result,
Parse would return an error because it thought the environment
variable was required. The new implementation fixes that
shortcoming and allows you to specify a default value of an
empty string.